+++ util.py
@@ -81,17 +81,24 @@ def make_prerequisits(project, folder, version_id, top_dir):
     raise Exception("Failed to build the dxWDL runtime asset")
 
 
-def build(project, folder, version_id, top_dir):
-    make_prerequisits(project, folder, version_id, top_dir)
-
+def find_asset(project, folder):
     # get asset_id
-    asset = dxpy.search.find_one_data_object(classname="record",
-                                             project=project.get_id(),
-                                             name="dxWDLrt",
-                                             folder=folder,
-                                             return_handler=True,
-                                             more_ok=False)
-    print("assetId={}".format(asset.get_id()))
+    assets = list(dxpy.search.find_data_objects(classname="record",
+                                                project=project.get_id(),
+                                                name="dxWDLrt",
+                                                folder=folder,
+                                                return_handler=True))
+    if len(assets) == 0:
+        return None
+    if len(assets) == 1:
+        return assets[0]
+    raise Exception("More than one asset found in folder {}".format(folder))
+
+def build(project, folder, version_id, top_dir):
+    asset = find_asset(project, folder)
+    if asset is None:
+        make_prerequisits(project, folder, version_id, top_dir)
+        asset = find_asset(project, folder)
 
     # update asset_id in configuration file
     top_conf_file = get_top_conf_file(top_dir)